refactor(DbtYamlManager): update how columns are retrieved
Update the get_columns method to account for the fact that in BigQuery Projects the columns can be nested structurs with their own documentation. This is accounted for by flattening these structs to make sure that they every name is accounted for. This logic will be used when the type returned by the adapter is a BigQuery Column that has the flatten method. Otherwise, the column names will be returned normally.